# src/api/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "klust3r API",
    "description": "klust3r APIs"
  },
  "basePath": "/",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/users/gameHint": {
      "get": {
        "tags": ["Game"],
        "summary": "Game hint  data",
        "description": "Game hint data",
        "operationId": "gameHint",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
          {
            "$ref": "#/parameters/HeaderAuth"
          }
          
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "responsesOk": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "msg": {
          "type": "string"
        },
        "data": {
          "type": "object"
        }
      }
    },
    "responsesError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "msg": {
          "type": "string"
        },
        "error": {
          "type": "object"
        }
      }
    },
...
  },
  "parameters": {
    "HeaderVersion": {
      "name": "X-klust3r-Version",
      "in": "header",
      "description": "X-klust3r-Version",
      "required": true,
      "type": "string",
      "default": "1.0.0"
    },
    "HeaderLanguage": {
      "name": "Accept-Language",
      "in": "header",
      "description": "Accept-Language",
      "required": true,
      "type": "string",
      "default": "en"
    },
    "HeaderPlatform": {
      "name": "X-klust3r-Platform",
      "in": "header",
      "description": "android or ios or web",
      "required": true,
      "type": "string",
      "default": "android"
    },
    "HeaderAuth": {
      "name": "Authorization",
      "in": "header",
      "description": "Authorization",
      "required": true,
      "type": "string"
    },
...
  }
}